+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.3.0 - Unreleased
 - added `global`and `nonlocal`statements
+- added single-line if/elif/else, for, while and def (e.g. `def f(x): return 42` without newline)
 
 ## 1.2.1
 - improve package description and fix linter warnings